+++ backend/logs/.0543e63d9e3fc74ae6dc8b3a2906b0062cebcd86-audit.json
@@ -14,6 +14,11 @@
             "date": 1749831503137,
             "name": "logs/combined-2025-06-13.log",
             "hash": "3439b7b167ee56e8a87124f87ba85c28628cb026bd6974650eba18f45e3588b9"
+        },
+        {
+            "date": 1749874794491,
+            "name": "logs/combined-2025-06-14.log",
+            "hash": "c0985b88973ed6520d0c12a2656bee0ee7686c87296ffa2c936c83177e48b7d6"
         }
     ],
     "hashType": "sha256"
